@@ -2,29 +2,46 @@
   "name": "@loke/http-rpc",
   "version": "3.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
     "fix": "eslint . --fix",
     "lint": "eslint .",
-    "test": "ava -v"
+    "test": "ava -v",
+    "prepare": "npm run build",
+    "build": "rm -rf ./dist && tsc"
   },
   "author": "Dominic Smith",
+  "ava": {
+    "extensions": [
+      "ts"
+    ],
+    "require": [
+      "ts-node/register"
+    ]
+  },
   "devDependencies": {
     "@loke/errors": "0.2.1",
-    "ava": "^0.15.2",
+    "@types/body-parser": "^1.19.0",
+    "@typescript-eslint/eslint-plugin": "^4.1.0",
+    "@typescript-eslint/parser": "^4.1.0",
+    "ava": "^3.12.1",
     "body-parser": "^1.15.2",
-    "eslint": "^5.9.0",
-    "eslint-config-prettier": "^3.3.0",
-    "eslint-plugin-prettier": "^3.0.0",
+    "eslint": "^7.8.1",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-ava": "^11.0.0",
+    "eslint-plugin-prettier": "^3.1.4",
     "express": "^4.14.0",
-    "got": "^6.3.0",
-    "prettier": "^1.15.2",
-    "prom-client": "^11.2.0"
+    "got": "^11.6.2",
+    "prettier": "2.1.1",
+    "prom-client": "^12.0.0",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.2"
   },
   "peerDependencies": {
-    "prom-client": "11.x"
+    "prom-client": "12.x"
   },
   "dependencies": {
-    "p-finally": "^1.0.0"
+    "@types/express": "^4.17.8"
   }
 }
